apollo-server-types: move info.cacheControl 'declare module' here
This makes `info.cacheControl` available to TypeScript packages that
depend on `apollo-server-types` rather than only being declared deep
inside `apollo-server-core`. Additionally, it gives a name to type used
for `info.cacheControl`.

Intended for use cases like
https://github.com/apollographql/federation/pull/870

We've run into tricky issues with `declare module` before so if this
ends up causing more problems than it's worth, we may revert it.
